refactor: removing unused and duplicated internal APIs
