+++ src/main/java/it/gov/pagopa/fdr/controller/interfaces/controller/IInternalOperationsController.java
@@ -1,0 +1,181 @@
+package it.gov.pagopa.fdr.controller.interfaces.controller;
+
+import it.gov.pagopa.fdr.controller.model.common.response.GenericResponse;
+import it.gov.pagopa.fdr.controller.model.flow.request.CreateFlowRequest;
+import it.gov.pagopa.fdr.controller.model.flow.response.SingleFlowCreatedResponse;
+import it.gov.pagopa.fdr.controller.model.payment.request.AddPaymentRequest;
+import it.gov.pagopa.fdr.controller.model.payment.request.DeletePaymentRequest;
+import it.gov.pagopa.fdr.util.constant.ControllerConstants;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Pattern;
+import jakarta.ws.rs.Consumes;
+import jakarta.ws.rs.DELETE;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.POST;
+import jakarta.ws.rs.PUT;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.PathParam;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.MediaType;
+import org.eclipse.microprofile.openapi.annotations.Operation;
+import org.eclipse.microprofile.openapi.annotations.media.Content;
+import org.eclipse.microprofile.openapi.annotations.media.Schema;
+import org.eclipse.microprofile.openapi.annotations.parameters.RequestBody;
+import org.eclipse.microprofile.openapi.annotations.responses.APIResponse;
+import org.eclipse.microprofile.openapi.annotations.responses.APIResponses;
+import org.eclipse.microprofile.openapi.annotations.tags.Tag;
+import org.jboss.resteasy.reactive.RestResponse;
+
+@Path(ControllerConstants.URL_CONTROLLER_INTERNAL_PSPS)
+@Consumes("application/json")
+@Produces("application/json")
+@Tag(name = "Internal Operations", description = "APIs for internal operations")
+public interface IInternalOperationsController {
+
+  @POST
+  @Path(ControllerConstants.URL_API_CREATE_EMPTY_FLOW)
+  @Operation(
+      operationId = "IInternalOperationsController.createEmptyFlowForInternalUse",
+      summary = "Create fdr",
+      description = "Create fdr")
+  @RequestBody(content = @Content(schema = @Schema(implementation = CreateFlowRequest.class)))
+  @APIResponses(
+      value = {
+        @APIResponse(ref = "#/components/responses/InternalServerError"),
+        @APIResponse(ref = "#/components/responses/AppException400"),
+        @APIResponse(ref = "#/components/responses/AppException404"),
+        @APIResponse(
+            responseCode = "201",
+            description = "Created",
+            content =
+                @Content(
+                    mediaType = MediaType.APPLICATION_JSON,
+                    schema = @Schema(implementation = GenericResponse.class)))
+      })
+  RestResponse<GenericResponse> createEmptyFlowForInternalUse(
+      @PathParam(ControllerConstants.PARAMETER_PSP) String pspId,
+      @PathParam(ControllerConstants.PARAMETER_FDR) @Pattern(regexp = "[a-zA-Z0-9\\-_]{1,35}")
+          String flowName,
+      @NotNull @Valid CreateFlowRequest request);
+
+  @PUT
+  @Path(ControllerConstants.URL_API_ADD_PAYMENT_IN_FLOW)
+  @Operation(
+      operationId = "IInternalOperationsController.addPaymentToExistingFlowForInternalUse",
+      summary = "Add payments to fdr",
+      description = "Add payments to fdr")
+  @RequestBody(content = @Content(schema = @Schema(implementation = AddPaymentRequest.class)))
+  @APIResponses(
+      value = {
+        @APIResponse(ref = "#/components/responses/InternalServerError"),
+        @APIResponse(ref = "#/components/responses/AppException400"),
+        @APIResponse(ref = "#/components/responses/AppException404"),
+        @APIResponse(
+            responseCode = "200",
+            description = "Success",
+            content =
+                @Content(
+                    mediaType = MediaType.APPLICATION_JSON,
+                    schema = @Schema(implementation = GenericResponse.class)))
+      })
+  GenericResponse addPaymentToExistingFlowForInternalUse(
+      @PathParam(ControllerConstants.PARAMETER_PSP) String pspId,
+      @PathParam(ControllerConstants.PARAMETER_FDR) String flowName,
+      @NotNull @Valid AddPaymentRequest request);
+
+  @PUT
+  @Path(ControllerConstants.URL_API_DELETE_PAYMENT_IN_FLOW)
+  @Operation(
+      operationId = "IInternalOperationsController.deletePaymentFromExistingFlowForInternalUse",
+      summary = "Delete payments to fdr",
+      description = "Delete payments to fdr")
+  @RequestBody(content = @Content(schema = @Schema(implementation = DeletePaymentRequest.class)))
+  @APIResponses(
+      value = {
+        @APIResponse(ref = "#/components/responses/InternalServerError"),
+        @APIResponse(ref = "#/components/responses/AppException400"),
+        @APIResponse(ref = "#/components/responses/AppException404"),
+        @APIResponse(
+            responseCode = "200",
+            description = "Success",
+            content =
+                @Content(
+                    mediaType = MediaType.APPLICATION_JSON,
+                    schema = @Schema(implementation = GenericResponse.class)))
+      })
+  GenericResponse deletePaymentFromExistingFlowForInternalUse(
+      @PathParam(ControllerConstants.PARAMETER_PSP) String pspId,
+      @PathParam(ControllerConstants.PARAMETER_FDR) String flowName,
+      @NotNull @Valid DeletePaymentRequest request);
+
+  @POST
+  @Path(ControllerConstants.URL_API_PUBLISH_FLOW)
+  @Operation(
+      operationId = "IInternalOperationsController.publishFlowForInternalUse",
+      summary = "Publish fdr",
+      description = "Publish fdr")
+  @APIResponses(
+      value = {
+        @APIResponse(ref = "#/components/responses/InternalServerError"),
+        @APIResponse(ref = "#/components/responses/AppException400"),
+        @APIResponse(ref = "#/components/responses/AppException404"),
+        @APIResponse(
+            responseCode = "200",
+            description = "Success",
+            content =
+                @Content(
+                    mediaType = MediaType.APPLICATION_JSON,
+                    schema = @Schema(implementation = GenericResponse.class)))
+      })
+  GenericResponse publishFlowForInternalUse(
+      @PathParam(ControllerConstants.PARAMETER_PSP) String pspId,
+      @PathParam(ControllerConstants.PARAMETER_FDR) String flowName);
+
+  @DELETE
+  @Path(ControllerConstants.URL_API_DELETE_FLOW)
+  @Operation(
+      operationId = "IInternalOperationsController.deleteExistingFlowForInternalUse",
+      summary = "Delete fdr",
+      description = "Delete fdr")
+  @APIResponses(
+      value = {
+        @APIResponse(ref = "#/components/responses/InternalServerError"),
+        @APIResponse(ref = "#/components/responses/AppException400"),
+        @APIResponse(ref = "#/components/responses/AppException404"),
+        @APIResponse(
+            responseCode = "200",
+            description = "Success",
+            content =
+                @Content(
+                    mediaType = MediaType.APPLICATION_JSON,
+                    schema = @Schema(implementation = GenericResponse.class)))
+      })
+  GenericResponse deleteExistingFlowForInternalUse(
+      @PathParam(ControllerConstants.PARAMETER_PSP) String pspId,
+      @PathParam(ControllerConstants.PARAMETER_FDR) String flowName);
+
+  @GET
+  @Path(ControllerConstants.URL_API_GET_SINGLE_NOT_PUBLISHED_FLOW)
+  @Operation(
+      operationId = "IInternalOperationsController.getSingleFlowNotInPublishedStatusForInternalUse",
+      summary = "Get created fdr",
+      description = "Get created fdr")
+  @APIResponses(
+      value = {
+        @APIResponse(ref = "#/components/responses/InternalServerError"),
+        @APIResponse(ref = "#/components/responses/AppException400"),
+        @APIResponse(ref = "#/components/responses/AppException404"),
+        @APIResponse(
+            responseCode = "200",
+            description = "Success",
+            content =
+                @Content(
+                    mediaType = MediaType.APPLICATION_JSON,
+                    schema = @Schema(implementation = SingleFlowCreatedResponse.class)))
+      })
+  SingleFlowCreatedResponse getSingleFlowNotInPublishedStatusForInternalUse(
+      @PathParam(ControllerConstants.PARAMETER_PSP) String pspId,
+      @PathParam(ControllerConstants.PARAMETER_FDR) String flowName,
+      @PathParam(ControllerConstants.PARAMETER_ORGANIZATION) String organizationId);
+}
